+++ src/plugins/Halftone.js
@@ -1,0 +1,203 @@
+/**
+*
+* Halftone Plugin
+* @package FILTER.js
+*
+**/
+!function(FILTER, undef){
+@@USE_STRICT@@
+
+var f1 = 7/16, f2 = 3/16, f3 = 5/16, f4 = 1/16, 
+    A32F = FILTER.Array32F, clamp = FILTER.Color.clamp,
+    RGB2YCbCr = FILTER.Color.RGB2YCbCr, YCbCr2RGB = FILTER.Color.YCbCr2RGB;
+
+// http://en.wikipedia.org/wiki/Halftone
+// http://en.wikipedia.org/wiki/Error_diffusion
+// http://www.visgraf.impa.br/Courses/ip00/proj/Dithering1/average_dithering.html
+// http://en.wikipedia.org/wiki/Floyd%E2%80%93Steinberg_dithering
+FILTER.Create({
+    name: "HalftoneFilter"
+    
+    // parameters
+    ,_threshold: 0.4
+    ,_size: 1
+    ,_grayscale: false
+    
+    // this is the filter constructor
+    ,init: function( size, threshold, grayscale ) {
+        var self = this;
+        self._size = size || 1;
+        self._threshold = clamp(undef === threshold ? 0.4 : threshold,0,1);
+        self._grayscale = !!grayscale;
+    }
+    
+    // support worker serialize/unserialize interface
+    ,path: FILTER.getPath( exports.AMD )
+    
+    ,size: function( s ) {
+        this._size = s;
+        return this;
+    }
+    
+    ,threshold: function( t ) {
+        this._threshold = clamp(t,0,1);
+        return this;
+    }
+    
+    ,grayscale: function( bool ) {
+        if ( !arguments.length ) bool = true;
+        this._grayscale = !!bool;
+        return this;
+    }
+    
+    ,serialize: function( ) {
+        var self = this;
+        return {
+            filter: self.name
+            ,_isOn: !!self._isOn
+            
+            ,params: {
+                _size: self._size
+                ,_threshold: self._threshold
+                ,_grayscale: self._grayscale
+            }
+        };
+    }
+    
+    ,unserialize: function( json ) {
+        var self = this, params;
+        if ( json && self.name === json.filter )
+        {
+            self._isOn = !!json._isOn;
+            
+            params = json.params;
+            
+            self._size = params._size;
+            self._threshold = params._threshold;
+            self._grayscale = params._grayscale;
+        }
+        return self;
+    }
+    
+    // this is the filter actual apply method routine
+    ,apply: function(im, w, h/*, image*/) {
+        var self = this, l = im.length, imSize = l>>>2,
+            err = new A32F(imSize*3), pixel, index, t, rgb, ycbcr,
+            size = self._size, area = size*size, invarea = 1.0/area,
+            threshold = 255*self._threshold, size2 = size2<<1,
+            colored = !self._grayscale,
+            x, y, yw, sw = size*w, i, j, jw, 
+            sum_r, sum_g, sum_b, qr, qg, qb
+            //,f11 = area*f1, f22 = area*f2, f33 = area*f3, f44 = area*f4
+        ;
+        
+        y=0; yw=0; x=0;
+        while ( y < h )
+        {
+            sum_r = sum_g = sum_b = 0;
+            i=0; j=0; jw=0;
+            while ( j < size )
+            {
+                pixel = (x+yw+i+jw)<<2; index = (x+yw+i+jw)*3;
+                sum_r += im[pixel] + err[index];
+                sum_g += im[pixel+1] + err[index+1];
+                sum_b += im[pixel+2] + err[index+2];
+                i++;
+                if ( i>=size ) {i=0; j++; jw+=w;}
+            }
+            sum_r *= invarea; sum_g *= invarea; sum_b *= invarea;
+            ycbcr = colored ? RGB2YCbCr([sum_r, sum_g, sum_b]) : [sum_r, sum_g, sum_b];
+            t = ycbcr[0];
+            if ( t > threshold )
+            {
+                if ( colored ) 
+                {
+                    ycbcr[0] = /*255;*/clamp(~~t, 0, 255);
+                    rgb = YCbCr2RGB(ycbcr);
+                }
+                else
+                {                    
+                    rgb = [255,255,255];
+                }
+            }
+            else
+            {                
+                rgb = [0,0,0];
+            }
+            pixel = (x+yw)<<2;
+            qr = im[pixel] - rgb[0];
+            qg = im[pixel+1] - rgb[1];
+            qb = im[pixel+2] - rgb[2];
+            
+            if ( x+size<w )
+            {                
+                i=size; j=0; jw=0;
+                while ( j < size )
+                {
+                    index = (x+yw+i+jw)*3;
+                    err[index] += f1*qr;
+                    err[index+1] += f1*qg;
+                    err[index+2] += f1*qb;
+                    i++;
+                    if ( i>=size2 ) {i=size; j++; jw+=w;}
+                }
+            }
+            if ( y+size<h && x>size) 
+            {
+                i=-size; j=size; jw=0;
+                while ( j < size2 )
+                {
+                    index = (x+yw+i+jw)*3;
+                    err[index] += f2*qr;
+                    err[index+1] += f2*qg;
+                    err[index+2] += f2*qb;
+                    i++;
+                    if ( i>=0 ) {i=-size; j++; jw+=w;}
+                }
+            }
+            if ( y+size<h ) 
+            {
+                i=0; j=size; jw=0;
+                while ( j < size2 )
+                {
+                    index = (x+yw+i+jw)*3;
+                    err[index] += f3*qr;
+                    err[index+1] += f3*qg;
+                    err[index+2] += f3*qb;
+                    i++;
+                    if ( i>=size ) {i=0; j++; jw+=w;}
+                }
+            }
+            if ( y+size<h && x+size<w )
+            {
+                i=size; j=size; jw=0;
+                while ( j < size2 )
+                {
+                    index = (x+yw+i+jw)*3;
+                    err[index] += f4*qr;
+                    err[index+1] += f4*qg;
+                    err[index+2] += f4*qb;
+                    i++;
+                    if ( i>=size2 ) {i=size; j++; jw+=w;}
+                }
+            }
+            
+            i=0; j=0; jw=0;
+            while ( j < size )
+            {
+                pixel = (x+yw+i+jw)<<2;
+                im[pixel] = rgb[0];
+                im[pixel+1] = rgb[1];
+                im[pixel+2] = rgb[2];
+                i++;
+                if ( i>=size ) {i=0; j++; jw+=w;}
+            }
+            
+            x+=size;
+            if ( x>=w ) {x=0; y+=size; yw+=sw;}
+        }
+        return im;
+    }
+});
+
+}(FILTER);