+++ changelog.md
@@ -8,6 +8,7 @@ __0.7__
 * add TGALoader, RGBELoader, GIFLoader generic/native image parsers/loaders
 * new plugin: **SeamlessTileFilter**, create a seamless tileable pattern from an image
 * new plugin: **HaarDetector**, detect image features using Viola-Jones-Lienhart openCV HAAR cascades algorithm (adapted from [HAAR.js](https://github.com/foo123/HAR.js))
+* new plugin: **HalftoneFilter**, create a halftoned/dithered image from target image
 * PerlinNoiseFilter plugin enhancement enable fractal noise and turbulence, octave noise and seamless noise
 * filters/plugins can transmit metadata (e.g HaarDetector) to pass additional information which is not necessarily an image
 * refactor/optimise